+++ CHANGES.md
@@ -15,6 +15,7 @@ New features:
 
 Internals:
 - Convert all materials to ShaderMaterial/RawShaderMaterial (#53)
+- Move some shaders to shader.js (#54)
 
 1.0.3
 -----